Split Slice.new into Slice.new and Slice.checked
Slice.checked performs an initial bounds check, while Slice.new just
uses its arguments as-is. This makes Slice.new useful for cases where
the range is already within bounds (e.g. inside a parsing loop), but the
compiler isn't able to figure that out.

In this setup Slice.new is still safe to use, as access to data of the
slice using e.g. Slice.get still performs a bounds check. Slice.pointer
is of course highly unsafe, but this was also the case prior to this
commit.