Perform basic bounds check in Slice.new
Instead of not performing any checks in `Slice.new`, we now perform a
simple check to assert that `start <= end` is always true. This is
needed as otherwise `Slice.size` might return a negative value, and
resolving that in `Slice.size` only complicates it more.

Due to the simplicity of the check it can be easily optimized away or
will be predicted by the CPU's branch predictor, such that the cost is
(or at least appears to be) close to zero, and much less compared to the
full slice bounds check performed by `Slice.checked`.