Fix checking type parameters as type arguments
When passing a generic type parameter as a type argument outside of
method bodies, the type parameter on the left (= originating from the
surrounding type) is a regular type parameter, yet the type it's
compared with may be a rigid version of that parameter. This ensures
that for such comparisons, we treat the type parameter on the left as
rigid.

Changelog: fixed