Use IndexMap/IndexSet where iteration is used
Ordering over HashMap/HashSet values isn't consistent. This ensures that
in all places where we do that, we now use an IndexMap/IndexSet, even if
this isn't strictly necessary. In addition, clippy is configured to
disallow iterating over HashMap/HashSet values.

The reason for this change is that the ordering in which data is
processed might affect incremental compilation, either by flushing
caches unnecessarily or by causing miscompilations (at least that has
been the case in the past).

Refer to https://github.com/inko-lang/inko/issues/808 for more details.

Changelog: changed