Use default<O2> for release builds
Instead of providing a custom (and slightly modified) list of LLVM
passes to run, we now just use the "default<O2>" group of passes. This
doesn't impact compile times and yields effectively the same executable,
yet it's a lot more portable compared to a custom list of passes that
has to be specialized for different LLVM versions.

This fixes https://github.com/inko-lang/inko/issues/871.

Changelog: changed