Use a dedicated exit status for test failures
This way we can detect the difference between a failure due to test
failures, and a failure due to a crash. This allows us to only show the
"test runner exited with ..." error message when necessary (i.e. not for
regular test failures).

This fixes https://github.com/inko-lang/inko/issues/860.

Changelog: fixed