Remove Entry from the public Map API
By making the Entry type a private type, and replacing its usage (e.g.
in Map.iter) with tuples, we can potentially optimize the Map internals
better in the future.

This change means the removal of the Map.entry API, though the way it
worked was a bit wonky and not that useful to begin with.

Refer to https://github.com/inko-lang/inko/issues/836 for more details.

Changelog: changed