Optimize the memory layout of Map
By storing the distance of each Entry in the "slots" array, we can make
Entry an inline type and avoid heap allocations for each entry. The
combination of these changes reduces the total amount of allocations,
and reduces the amount of memory necessary (e.g. due to Entry no longer
having a header).

In addition, Map.with_capacity(N) is adjusted to ensure there is indeed
enough space for N elements before a resize, whereas previously it would
still apply the load factor to it, meaning the actual number of pairs
one could insert before a resize would be 90% of N. Map.capacity in turn
simply reports the "resize_at" threshold, which correctly reflects the
number of pairs that can be inserted before a resize.

This fixes https://github.com/inko-lang/inko/issues/836.

Changelog: performance