Use a more compact layout for std.map.Slot
By using two i32 values we can reduce the size of a single Slot from 16
bytes to just 8 bytes. The trade-off is that Map is now limited to a
maximum of 1 932 735 283 key-value pairs. To ensure such maps don't
trigger undefined behavior, Map.resize panics when exceeding this limit.
In reality it's _highly_ unlikely (read: more or less impossible) to
reach this limit to begin with, so it shouldn't cause any problems.