Remove pooling of process stacks
The StackPool implementation was such that a program could use large
amounts of memory when spawning many processes, with most of that memory
seemingly residing in kernel space.

To solve this, we remove the StackPool implementation entirely and
allocate and set up process stacks when they run their first message.
This allows hiding of the latency from the spawning/creating
process/thread, and removes the need for allocating additional memory
for managing stacks. This in turn ensures that memory usage remains
small and stable, even when spawning many processes.

With this change the time it takes to allocate a process is now roughly
1 nanosecond, as it amounts to a function call and a malloc(). Sending
messages of course still comes with a (greater) cost, but that's a
separate subject.

This fixes https://github.com/inko-lang/inko/issues/875.

Changelog: performance