Add support for sending of files to sockets
This adds a `send_file` method to various socket types, as provided by
the `SendFile` trait. The `send_file` method is used to efficiently send
a file to a socket, without the need for copying its data in userspace.
This method uses the `sendfile()` system call on the currently supported
platforms. For datagram sockets on FreeBSD and macOS we use a userspace
fallback.

For TLS sockets we fall back to performing a userspace copy as we don't
support offloading TLS work to the kernel. This is implemented using the
new `std.io.copy` method which copies data from a `Read` value to a
`Write` value in chunks.

This commit also includes some changes to the `Seek` trait such that
it's now possible to also request the current cursor position using
`std.io.SeekFrom.Current(0)`.

Changelog: added