Fix use after free when configuring TLS clients
When configuring a TLS client a server name is required, used for SNI.
The ServerName type from rustls supports both T and &T as input. As we
passed a &str it ended up _borrowing_ the data, but the data wouldn't
necessarily outlive the TLS client.

To fix this, the server name is explicitly cloned into a Rust String
before it's passed to rustls, ensuring rustls owns the data for as long
as necessary.

This fixes https://github.com/inko-lang/inko/issues/886

Changelog: fixed