Resolve fields and methods with the same name
If a type defines a field `foo` and a method `foo`, we now support
referring to the field when not using any arguments and parentheses,
instead of such an expression always resolving to a method call. This
means you can have (for example) a field `status` and a method
`status(arg)` and use `x.status` to refer to the field and `x.status(n)`
to call the method.

Changelog: added