Fix accept() getting stuck for Unix sockets
When cloning an IP socket into sockets A and B, if you call `accept(A)`
followed by `shutdown(B)`, then `accept(A)` returns with EINVAL.
However, if the sockets are Unix sockets then `accept(A)` keeps
returning EAGAIN. This would then result in `TcpServer.accept` getting
stuck in a loop trying to call `accept()` followed by polling the socket
over and over again.

To fix this we only allow up to two attempts: one for the initial
`accept()` that might block (and is thus followed by a poll), and a
second attempt after waking up from the poll. If this second attempt
still fails, it must be because the socket is closed.

Changelog: fixed