Remove accept return error normalization
I was wrong about accept() returning ENOTCONN, that was caused by the
shutdown() call in the tests. On macOS and FreeBSD it seems you can't
actually interrupt accept() with shutdown() at all (as in, it's Linux
specific). I guess that's what happens when you're still coding in the
middle of the night.