Add basic MIME type detection
This adds the module std.mime which provides a Mime type. This type can
be used to detect MIME types from file extensions, and possibly more in
the future (e.g. by sniffing file contents).

The implementation relies on a perfect hash function generated ahead of
time, based on a JSON file containing a mapping of file extensions to
their MIME types.

This commit also adds the methods `String.to_slice`,
`ByteArray.to_slice` and `Slice.to_slice`, provided by the `ToSlice`
trait. This makes it easier for methods to operate on both `String` and
`Slice[String]` values, something the new `Mime` type makes use of to
allow parsing of both such values.

This differs from using a `Bytes` as input in that it allows you to take
advantage of `String` being reference counted, removing the need for any
cloning and the heap allocations this may require. For 98% of the other
use cases the `Bytes` trait is better suited.

Changelog: added