Replace Bool.true?/false? with an operator
The methods `Bool.true?` and `Bool.false?` were a remnant from when Inko
was closer to Smalltalk (e.g. `if` being a method instead of an
expression). While using methods makes the syntax simpler, it makes code
more annoying to read as the condition (inversion) is now at the end,
meaning you have to read it in reverse. That is, this:

    if foo.false? { ... }

Reads out as "if foo is false" or "if foo is not" or something along
those lines. This is OK for short conditions, but annoying for longer
ones. In contrast, this:

    if !foo { ... }

Reads out as "if not foo", which works much better for longer
conditions.

This commit introduces support for the syntax `!expr` and removes the
`Bool.true?` and `Bool.false?` methods in favor of this new operator.

Changelog: changed