Don't add module dependencies for inlined code
This setup wasn't working correctly and would result in too many caches
being flushed without this being necessary. For example, refreshing
std.net.http.server would also refresh more or less the entire standard
library.

In addition, it's redundant: for a module to use some inlined type or
method it must either directly or indirectly depend on it through an
import. This means that a change to the original source module
eventually trickles down to the module the code is inlined into, without
the need for additional tracking.

We do still track the method IDs that are inlined as that _is_ still
necessary.

Changelog: performance