Add modules for generating XML and HTML documents
This adds the modules `std.xml` and `std.html` for generating XML and
HTML respectively. The API is based on that of
https://github.com/yorickpeterse/inko-builder, but without the need for
creating an intermediate DOM tree. This allows for faster and more
memory efficient generating of documents, at the cost of not being able
to modify them after they're generated (at least not without a fully
fledged XML or HTML parser).

Changelog: added