Add the ability to dump MIR to text files
While the DOT format is useful for generating graphs of simple methods,
it doesn't work well for larger ones (e.g. it takes forever to generate
an SVG). In addition, navigating around dot output isn't exactly a nice
experience.

This commit adds the ability to dump MIR to its text format, and cleans
up the DOT dumping a bit as well.

Changelog: added