Generate better code for match expressions
The code generated for match expressions with enums was horribly
inefficient. This was due to the pattern matching compiler (essentially)
generating a branch of each enum constructor, even if only a select few
were specified as a pattern. This could result in an explosion of
branches when comparing enums as part of e.g. a tuple pattern. For
example:

    match (a, b) {
      case (A, B) -> true
      case _ -> false
    }

Assuming `a` and `b` are enums with each having 100 constructors, this
code would generate 100 branches for the first value in the tuple, and
then 100 branches for the second value _for each_ first branch. In other
words, the total number of branches would be at least 100 * 100.
Depending on the exact type this could further explode by inlining their
increment/decrement/drop glue methods, potentially resulting in hundreds
of thousands basic blocks/branches.

While our own compiler logic didn't care too much about it, LLVM
_really_ doesn't like compiling a function with 500 000 basic blocks and
could easily spend 30 seconds doing so.

This commit optimizes this by essentially only generating branches for
enum constructors actually specified as a pattern, delegating the others
to a fallback branch. In addition we run some optimizations on MIR to
merge certain basic blocks and further compress the switch()
instructions generated by pattern matching logic.

The result of these changes is a dramatic improvement in compile times
of match expressions that match against large enums.

This fixes https://github.com/inko-lang/inko/issues/906.

Changelog: performance