Add async and await expressions
In the past, sending a message to a process would result in the sender
waiting for the response immediately, unless opted out using the `async`
keyword in which case a Future was returned.

This commit sort of reintroduces this, but in a way that composes
better. Sending messages behaves as before, i.e. the sender _doesn't_
wait, and a `Promise` must be used to communicate the result back to the
sender. To reduce the boilerplate that comes with using `Future.new`,
`async` and `await` _expression_ may be used. These expressions are just
syntax sugar for pattern matching. For example, this:

    async foo

Compiles to the following:

    match Future.new {
      case (future, promise) -> {
        foo(promise)
        future
      }
    }

And this:

    await foo

Compiles to this:

    match Future.new {
      case (future, promise) -> {
        foo(promise)
        future.get
      }
    }

Using these expressions it's much easier to send message and wait for
their responses, while still having full control over how `Promise` and
`Future` values are passed around (e.g. you can still manually pass a
`Promise` if desired).

Changelog: added