Add support for pattern matching in `let`
The `let` expression now fully supports pattern matching, instead of
being limited to simply defining a binding with a value. If the pattern
is not exhaustive/can fail, an `else` expression is required that
diverges from the surrounding method. For example:

    let Some(value) = something else return 42

This reduces the amount of nesting that's necessary when destructuring
certain values such as tuples. For example, instead of this:

    match some_tuple {
      case (a, b) -> {
        ...
      }
    }

You can now write this instead:

    let (a, b) = some_tuple

And instead of this:

    let val = match Int.parse(input, Format.Decimal) {
      case Some(v) -> v
      case _ -> return ...
    }

You can write this:

    let Some(val) = Int.parse(input, Format.Decimal) else return ...

This fixes https://github.com/inko-lang/inko/issues/558.

Changelog: added