Ensure type signatures correctly create borrows
If a type signature contained a borrow such as `mut String` or `ref
SomeCopyType`, we'd construct a `TypeRef::Mut(...)` or
`TypeRef::Ref(...)` directly, instead of going through methods such as
`TypeRef::as_mut()`. This meant that for value types we'd use the wrong
form of ownership (i.e.g `TypeRef::Mut` instead of `TypeRef::Owned`).
This in turn could result in certain correct code producing a
compile-time error, such as the one seen in
https://github.com/inko-lang/inko/issues/876.

This commit fixes this issue and adjusts the compiler to produce a
warning for redundant borrows of value types (except for pointers).

Changelog: fixed