Fix hashing of Int values
The implementation of the `SipHash13` type is such that it expects a
_byte_ as input, not an arbitrary 64-bits integer. This means that when
given values outside of the inclusive range 0..255, the resulting hash
code degrades severely as inputs are essentially truncated.

This commit fixes this by breaking the `Hash` type into two methods:

- `Hash.write_byte`
- `Hash.write_int`

The first method is the old `Hash.write` method and is expected to be
given a value in the byte range. We don't check for this at runtime due
to the extra cost this incurs, but in the future we may further refactor
this should we ever introduce a dedicated byte type (see
https://github.com/inko-lang/inko/issues/834 for some more thoughts and
details on the subject).

The second method is meant to be given a regular `Int` of any value. In
the case of `SipHasher13` it's optimized such that each byte in the
`Int` is taken into consideration, but without the need for calling
`write_byte` for each said byte.

This fixes https://github.com/inko-lang/inko/issues/923.

Changelog: performance