Use sendfile for HTTP servers where possible
This changes the Response.write method such that whole file responses
use sendfile() (or the equivalent system call) when possible. While this
isn't terribly useful for TLS as we don't have kernel TLS support, and
useless for HTTP 2 and 3, it can still be useful when an Inko HTTP
application sits behind an HTTP(s) proxy that _does_ support kernel TLS,
or if regular HTTP is used.

As part of this the SendFile trait's API is changed to also take a
buffer so that it can be reused, and the userspace copy size is
increased to 256 KiB to reduce the amount of reads and writes.

Changelog: changed