Remove unused sections/functions during linking
Linkers have an option to remove unused symbols/sections (= functions)
during linking, such as the `--gc-sections` flag for the various
ld-compatible linkers. For this to work well, each function should have
its own section in the assembly/object file.

We now pass the correct option so the linker is able to remove unused
functions from the Rust runtime library. We don't generate unique
section files for Inko methods for the following reasons:

1. We have our own dead code elimination pass that should be able to
   remove unused methods, sans those that _might_ be called through
   dynamic dispatch
2. We don't support creating that can be linked against by other
   languages
3. LLVM's C++ API has support for automatically generating section
   names, but this isn't exposed through the C API and thus not
   available to Inkwell, so we'd have to generate our own names
4. For macOS I ran into some issues with it spitting out many warnings,
   and triggering assertion failures in ld64. I wasn't able to figure
   out why, nor was I able to figure out who else might've run into
   this
5. For macOS there's a limit of 16 characters for the entire section
   name, which seems to include the "__TEXT,__" prefix, meaning you
   actually only have 8 characters (= 64 bits) available. This isn't too
   bad, but may result in collisions and it's unclear what happens then

Rather than keep fighting this, I just went with the simplest option
which is just enabling the linker flag. This however still works quite
well: for a simple Hello World program this can reduce the executable
size by 50-60%, though this difference (obviously) becomes smaller the
more of the standard library (and thus the runtime library) is used.

Changelog: performance

fix