Add support for gzip compression and decompression
This adds the module std.compress.gzip, providing the types Encoder,
Decoder and Reader for compressing and decompressing data using the gzip
compression format.

The implementation is backed by the zlib-rs crate [1], removing the need
for installing zlib(-ng) in the build and target environments. This also
makes cross-compilation easier. The performance of zlib-rs should be on
par with that of zlib-ng, which is to say it should be good enough for
most people.

This commit also includes some changes related to the CLI and its HTTP
logic. While this is unrelated to the addition of gzip support, the
introduction of zlib-rs changed the dependency graph such that
cargo-deny started producing warnings about duplicate dependencies. To
resolve these the CLI now configures ureq to _not_ install the ring
crate any more and instead to also use graviola [2]. The version of
graviola remains the same as newer versions appear to introduce
compile-time regressions [3] that we likely won't benefit from in any
way.

[1]: https://github.com/trifectatechfoundation/zlib-rs
[2]: https://github.com/ctz/graviola
[3]: https://github.com/ctz/graviola/issues/114

Changelog: added