Handle borrows when checking for sendable returns
When checking if a return value is sendable, we already handled fields
being assigned to regular type parameters but not borrows of type
parameters. This in turn meant that e.g. `Slice[String]` wouldn't be
considered sendable, even though it is (as `ref String` is `String`
which is a value type).

Changelog: fixed