Use shutdown to interrupt accept for HTTP servers
This changes the graceful shutdown handling of
std.net.http.server.Server to use shutdown() to interrupt the server's
accept() loop. In an early implementation we did just that but found
this didn't work reliably.

Upon taking a closer look as part of
https://github.com/inko-lang/inko/issues/937, it was discovered that
it's only macOS that doesn't support this pattern, with both Linux and
FreeBSD supporting it just fine, and with OpenBSD also supporting it
(though we currently don't support OpenBSD).

This commit changes the Server implementation to use the shutdown
pattern (again) as it makes for more reliable shutdowns when multiple
instances of the same server are running.

This fixes https://github.com/inko-lang/inko/issues/937.

Changelog: changed