Make interrupting accept calls portable
On Linux, FreeBSD, OpenBSD and possibly other BSDs one can interrupt a
call to `TcpServer.accept` by cloning the socket using
`TcpServer.try_clone` and then calling `shutdown` on the socket,
shutting down the socket and marking it as readable such that epoll/etc
can be used to reschedule the process.

Unfortunately macOS is a bit of a party pooper and doesn't support this
_at all_: calling `shutdown()` on a listening socket produces `ENOTCONN`
because the socket isn't connected to anything, and no shutdown logic is
triggered.

Using `close()` instead of `shutdown()` doesn't work either, because if
simply removes the file descriptor from the epoll/kqueue instances it's
registered with without any sort of notification.

The result is that the only portable way of interrupting an `accept`
call asynchronously is to use an extra IO object of sorts which you then
poll in _addition_ to a socket _before_ calling `accept()`. You can then
write to/signal this extra IO object such that it becomes available for
reading, waking up the network poller.

In this commit we do exactly that, using a pipe as the extra IO object.
While more lightweight options exist such as eventfd on Linux and
FreeBSD or using kqueue with `EVFILT_USER` events, pipes are portable
across all platforms and were sort of already available (but hidden) in
the `std.sys` module as they're used for communication with child OS
processes. The cost of pipes isn't really relevant here either because
realistically you'll only have a few socket listeners per program, and
shutdown signals are only triggered once.

This requires some changes to the scheduler to support polling two event
sources at the same time instead of just one. While I'm not super happy
with the resulting code, it was the least messy solution I could come up
with. Crucially, it doesn't require implementing epoll/kqueue wrappers
in the standard library for _just_ this use case.

As part of these changes, the standard library now supports pipes
through the `std.pipe` module and the `std.net.http.Server` type is
adjusted to use this new setup for gracefully shutting down HTTP
servers.

This fixes https://github.com/inko-lang/inko/issues/937.

Changelog: fixed