Remove atomics for updating poll registrations
When updating the "registered" state of a pollable source, atomics
shouldn't be necessary as we already acquire the process' run lock and
don't release it until registration is finished. This prevents the same
process from running concurrently on different threads, and should also
ensure all threads observe a consistent view of the source's state.

Of course this being atomics, I may one day be proven wrong :)

Changelog: changed