Fix handling of certain signals across platforms
Certain signals such as SIGURG and SIGCONT are discarded entirely
instead of just not doing anything. Most notably macOS does this, while
Linux and FreeBSD seem to just do nothing but still deliver the signals
to the process. To resolve this we need to install a noop handler for
these signals.

In addition this commit also refactors std.signal to use the correct
signal values across platforms, and we add the ability to send signals
to processes started using `std.sys.Command`.

This fixes https://github.com/inko-lang/inko/issues/943

Changelog: fixed